+++ tests/Unit/Infrastructure/Rule/Factories/RuleFactoryTest.php
@@ -6,14 +6,14 @@ use ArtARTs36\MergeRequestLinter\Application\Rule\Rules\AbstractRule;
 use ArtARTs36\MergeRequestLinter\Domain\Request\MergeRequest;
 use ArtARTs36\MergeRequestLinter\Domain\Rule\RuleDefinition;
 use ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Argument\Builder;
-use ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Argument\Resolvers\AsIsResolver;
-use ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Argument\Resolvers\CompositeResolver;
-use ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Argument\Resolvers\MapResolver;
-use ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Argument\Resolvers\ObjectCompositeResolver;
 use ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Factories\RuleFactory;
 use ArtARTs36\MergeRequestLinter\Shared\DataStructure\ArrayMap;
 use ArtARTs36\MergeRequestLinter\Shared\DataStructure\Map;
 use ArtARTs36\MergeRequestLinter\Shared\Reflection\Instantiator\Finder;
+use ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver\AsIsResolver;
+use ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver\CompositeResolver;
+use ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver\MapResolver;
+use ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver\ObjectCompositeResolver;
 use ArtARTs36\MergeRequestLinter\Tests\TestCase;
 
 final class RuleFactoryTest extends TestCase
