+++ tests/Unit/Shared/Reflection/TypeResolver/ArrayeeResolverTest.php
@@ -1,21 +1,21 @@
 <?php
 
-namespace ArtARTs36\MergeRequestLinter\Tests\Unit\Infrastructure\Rule\Argument\Resolvers;
+namespace ArtARTs36\MergeRequestLinter\Tests\Unit\Shared\Reflection\TypeResolver;
 
-use ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Argument\Resolvers\SetResolver;
 use ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Exceptions\ArgNotSupportedException;
-use ArtARTs36\MergeRequestLinter\Shared\DataStructure\Set;
+use ArtARTs36\MergeRequestLinter\Shared\DataStructure\Arrayee;
 use ArtARTs36\MergeRequestLinter\Shared\Reflection\Reflector\Type;
 use ArtARTs36\MergeRequestLinter\Shared\Reflection\Reflector\TypeName;
+use ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver\ArrayeeResolver;
 use ArtARTs36\MergeRequestLinter\Tests\TestCase;
 
-final class SetResolverTest extends TestCase
+final class ArrayeeResolverTest extends TestCase
 {
     public function providerForTestCanResolve(): array
     {
         return [
             [
-                new Type(TypeName::Object, Set::class),
+                new Type(TypeName::Object, Arrayee::class),
                 [],
                 true,
             ],
@@ -28,12 +28,12 @@ final class SetResolverTest extends TestCase
     }
 
     /**
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Argument\Resolvers\SetResolver::canResolve
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver\ArrayeeResolver::canResolve
      * @dataProvider providerForTestCanResolve
      */
     public function testCanResolve(Type $type, mixed $value, bool $expected): void
     {
-        $resolver = new SetResolver();
+        $resolver = new ArrayeeResolver();
 
         self::assertEquals(
             $expected,
@@ -44,36 +44,33 @@ final class SetResolverTest extends TestCase
     public function providerForTestResolve(): array
     {
         return [
-            [
-                new Type(TypeName::Array),
-                ['value1'],
-            ],
+            [new Type(TypeName::Array), ['val1']],
         ];
     }
 
     /**
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Argument\Resolvers\SetResolver::resolve
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver\ArrayeeResolver::resolve
      * @dataProvider providerForTestResolve
      */
     public function testResolve(Type $type, mixed $value): void
     {
-        $resolver = new SetResolver();
+        $resolver = new ArrayeeResolver();
 
         $result = $resolver->resolve($type, $value);
 
-        self::assertInstanceOf(Set::class, $result);
-        self::assertEquals($value, $result->values());
+        self::assertInstanceOf(Arrayee::class, $result);
+        self::assertEquals($value, $result->mapToArray(fn ($val) => $val));
     }
 
     /**
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Argument\Resolvers\SetResolver::resolve
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver\ArrayeeResolver::resolve
      */
-    public function testResolveOnArgNotSupportedException(): void
+    public function testResolveOnArgNotSupported(): void
     {
-        $resolver = new SetResolver();
+        $resolver = new ArrayeeResolver();
 
         self::expectException(ArgNotSupportedException::class);
 
-        $resolver->resolve(new Type(TypeName::Array), 1);
+        $resolver->resolve(new Type(TypeName::Array), 'string');
     }
 }
