+++ tests/Unit/Shared/Reflection/TypeResolver/EnumResolverTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace ArtARTs36\MergeRequestLinter\Tests\Unit\Infrastructure\Rule\Argument\Resolvers;
+namespace ArtARTs36\MergeRequestLinter\Tests\Unit\Shared\Reflection\TypeResolver;
 
-use ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Argument\Resolvers\EnumResolver;
 use ArtARTs36\MergeRequestLinter\Shared\Reflection\Reflector\Type;
 use ArtARTs36\MergeRequestLinter\Shared\Reflection\Reflector\TypeName;
+use ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver\EnumResolver;
 use ArtARTs36\MergeRequestLinter\Tests\TestCase;
 
 final class EnumResolverTest extends TestCase
@@ -23,7 +23,7 @@ final class EnumResolverTest extends TestCase
                 false,
             ],
             [
-                new Type(TypeName::String, TestEnum::class),
+                new Type(TypeName::String, \ArtARTs36\MergeRequestLinter\Tests\Unit\Infrastructure\Rule\Argument\Resolvers\TestEnum::class),
                 '',
                 true,
             ],
@@ -31,7 +31,7 @@ final class EnumResolverTest extends TestCase
     }
 
     /**
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Argument\Resolvers\EnumResolver::canResolve
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver\EnumResolver::canResolve
      * @dataProvider providerForTestCanResolve
      */
     public function testCanResolve(Type $type, mixed $value, bool $expected): void
@@ -55,20 +55,20 @@ final class EnumResolverTest extends TestCase
                 'Type with name "non-enum" not supported',
             ],
             [
-                new Type(TypeName::String, TestEnum::class),
+                new Type(TypeName::String, \ArtARTs36\MergeRequestLinter\Tests\Unit\Infrastructure\Rule\Argument\Resolvers\TestEnum::class),
                 new \stdClass(),
-                'Value for enum '. TestEnum::class .' must be int',
+                'Value for enum '. \ArtARTs36\MergeRequestLinter\Tests\Unit\Infrastructure\Rule\Argument\Resolvers\TestEnum::class .' must be int',
             ],
             [
-                new Type(TypeName::Int, TestEnum::class),
+                new Type(TypeName::Int, \ArtARTs36\MergeRequestLinter\Tests\Unit\Infrastructure\Rule\Argument\Resolvers\TestEnum::class),
                 4,
-                'Enum "'. TestEnum::class .'" not resolved. Available values: [1, 2, 3]',
+                'Enum "'. \ArtARTs36\MergeRequestLinter\Tests\Unit\Infrastructure\Rule\Argument\Resolvers\TestEnum::class .'" not resolved. Available values: [1, 2, 3]',
             ],
         ];
     }
 
     /**
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Argument\Resolvers\EnumResolver::resolve
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver\EnumResolver::resolve
      * @dataProvider providerForTestResolveOnException
      */
     public function testResolveOnException(Type $type, mixed $value, string $exception): void
@@ -81,14 +81,14 @@ final class EnumResolverTest extends TestCase
     }
 
     /**
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Argument\Resolvers\EnumResolver::resolve
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver\EnumResolver::resolve
      */
     public function testResolve(): void
     {
         $resolver = new EnumResolver();
 
-        self::assertEquals(TestEnum::Val1, $resolver->resolve(
-            new Type(TypeName::Int, TestEnum::class),
+        self::assertEquals(\ArtARTs36\MergeRequestLinter\Tests\Unit\Infrastructure\Rule\Argument\Resolvers\TestEnum::Val1, $resolver->resolve(
+            new Type(TypeName::Int, \ArtARTs36\MergeRequestLinter\Tests\Unit\Infrastructure\Rule\Argument\Resolvers\TestEnum::class),
             1,
         ));
     }
