+++ tests/Unit/Shared/Reflection/TypeResolver/SetResolverTest.php
@@ -1,21 +1,21 @@
 <?php
 
-namespace ArtARTs36\MergeRequestLinter\Tests\Unit\Infrastructure\Rule\Argument\Resolvers;
+namespace ArtARTs36\MergeRequestLinter\Tests\Unit\Shared\Reflection\TypeResolver;
 
-use ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Argument\Resolvers\ArrayeeResolver;
 use ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Exceptions\ArgNotSupportedException;
-use ArtARTs36\MergeRequestLinter\Shared\DataStructure\Arrayee;
+use ArtARTs36\MergeRequestLinter\Shared\DataStructure\Set;
 use ArtARTs36\MergeRequestLinter\Shared\Reflection\Reflector\Type;
 use ArtARTs36\MergeRequestLinter\Shared\Reflection\Reflector\TypeName;
+use ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver\SetResolver;
 use ArtARTs36\MergeRequestLinter\Tests\TestCase;
 
-final class ArrayeeResolverTest extends TestCase
+final class SetResolverTest extends TestCase
 {
     public function providerForTestCanResolve(): array
     {
         return [
             [
-                new Type(TypeName::Object, Arrayee::class),
+                new Type(TypeName::Object, Set::class),
                 [],
                 true,
             ],
@@ -28,12 +28,12 @@ final class ArrayeeResolverTest extends TestCase
     }
 
     /**
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Argument\Resolvers\ArrayeeResolver::canResolve
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver\SetResolver::canResolve
      * @dataProvider providerForTestCanResolve
      */
     public function testCanResolve(Type $type, mixed $value, bool $expected): void
     {
-        $resolver = new ArrayeeResolver();
+        $resolver = new SetResolver();
 
         self::assertEquals(
             $expected,
@@ -44,33 +44,36 @@ final class ArrayeeResolverTest extends TestCase
     public function providerForTestResolve(): array
     {
         return [
-            [new Type(TypeName::Array), ['val1']],
+            [
+                new Type(TypeName::Array),
+                ['value1'],
+            ],
         ];
     }
 
     /**
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Argument\Resolvers\ArrayeeResolver::resolve
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver\SetResolver::resolve
      * @dataProvider providerForTestResolve
      */
     public function testResolve(Type $type, mixed $value): void
     {
-        $resolver = new ArrayeeResolver();
+        $resolver = new SetResolver();
 
         $result = $resolver->resolve($type, $value);
 
-        self::assertInstanceOf(Arrayee::class, $result);
-        self::assertEquals($value, $result->mapToArray(fn ($val) => $val));
+        self::assertInstanceOf(Set::class, $result);
+        self::assertEquals($value, $result->values());
     }
 
     /**
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Argument\Resolvers\ArrayeeResolver::resolve
+     * @covers \ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver\SetResolver::resolve
      */
-    public function testResolveOnArgNotSupported(): void
+    public function testResolveOnArgNotSupportedException(): void
     {
-        $resolver = new ArrayeeResolver();
+        $resolver = new SetResolver();
 
         self::expectException(ArgNotSupportedException::class);
 
-        $resolver->resolve(new Type(TypeName::Array), 'string');
+        $resolver->resolve(new Type(TypeName::Array), 1);
     }
 }
